Move var/spool/cwl detector to process.py and make it a warning.
Do the check alongside other static checks.

Need to decide how best to fix the test that previously assumed this
was a validation error.
